refactor: split blockchain.rs into focused modules (Step 2)
Pure refactor — zero logic change. Splits blockchain.rs (1665 lines) into 6
focused modules:
- blockchain.rs  — struct, constants, genesis, core state methods (~170 lines)
- mempool.rs     — add_to_mempool, prune_mempool, mempool queries
- block_producer.rs — create_block
- block_executor.rs — add_block (two-pass validation + commit)
- token_ops.rs   — deploy_token, token_transfer, token_burn, token_balance, list_tokens
- chain_queries.rs  — get_transaction, get_address_history, chain_stats, richlist

Also includes Step 1 enforcement (cargo-deny + clippy -D warnings) from
ci: add cargo-deny + clippy enforcement (Step 1), consolidated here since
Step 1 PR is still open on GitHub.

186 tests pass, 0 clippy warnings, cargo deny ok.
