+++ src/core/account.rs
@@ -85,7 +85,7 @@ impl AccountDB {
         self.credit(to, amount)?;
 
         // L-02 FIX: Burn rounds up (ceiling) so odd fees are never lost — validator gets floor
-        let burn_amount = (fee + 1) / 2;
+        let burn_amount = fee.div_ceil(2);
         self.total_burned = self.total_burned.saturating_add(burn_amount);
 
         Ok(())
