+++ src/core/blockchain.rs
@@ -1,9 +1,9 @@
-// blockchain.rs - Sentrix
+// blockchain.rs - Sentrix — Blockchain struct, constants, genesis, core state methods
 
 use serde::{Deserialize, Serialize};
 use std::collections::VecDeque;
 use crate::core::block::Block;
-use crate::core::transaction::{Transaction, TokenOp};
+use crate::core::transaction::Transaction;
 use crate::core::account::AccountDB;
 use crate::core::authority::AuthorityManager;
 use crate::core::merkle::merkle_root;
@@ -141,114 +141,6 @@ impl Blockchain {
         reward.min(remaining)
     }
 
-    // ── Mempool ──────────────────────────────────────────
-    pub fn add_to_mempool(&mut self, tx: Transaction) -> SentrixResult<()> {
-        if tx.is_coinbase() {
-            return Err(SentrixError::InvalidTransaction(
-                "cannot manually add coinbase to mempool".to_string()
-            ));
-        }
-
-        // C-03 FIX: Global mempool size limit
-        if self.mempool.len() >= MAX_MEMPOOL_SIZE {
-            return Err(SentrixError::InvalidTransaction(
-                "mempool full — try again later".to_string()
-            ));
-        }
-
-        // C-03 FIX: Per-sender pending tx limit
-        let sender_pending = self.mempool_pending_count(&tx.from_address) as usize;
-        if sender_pending >= MAX_MEMPOOL_PER_SENDER {
-            return Err(SentrixError::InvalidTransaction(
-                "too many pending transactions from this sender".to_string()
-            ));
-        }
-
-        // H-04 FIX: Validate to_address is a well-formed Sentrix address
-        if !is_valid_sentrix_address(&tx.to_address) {
-            return Err(SentrixError::InvalidTransaction(
-                format!("invalid to_address: '{}'", tx.to_address)
-            ));
-        }
-
-        // M-03/M-04 FIX: Validate transaction timestamp
-        // Reject timestamps too far in the future (clock skew / pre-signed attack)
-        // Reject timestamps too old (replay of stale transactions / mempool poisoning)
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        if tx.timestamp > now + 300 {
-            return Err(SentrixError::InvalidTransaction(
-                "transaction timestamp too far in the future (max +5 min)".to_string()
-            ));
-        }
-        if now > tx.timestamp.saturating_add(MEMPOOL_MAX_AGE_SECS) {
-            return Err(SentrixError::InvalidTransaction(
-                format!("transaction too old — max age {} seconds", MEMPOOL_MAX_AGE_SECS)
-            ));
-        }
-
-        // Basic validation
-        let expected_nonce = self.accounts.get_nonce(&tx.from_address)
-            + self.mempool_pending_count(&tx.from_address);
-        tx.validate(expected_nonce, self.chain_id)?;
-
-        // Check balance including pending mempool spends
-        let pending_spend = self.mempool_pending_spend(&tx.from_address);
-        let available = self.accounts.get_balance(&tx.from_address)
-            .saturating_sub(pending_spend);
-        // H-02 FIX: checked addition to prevent overflow
-        let needed = tx.amount.checked_add(tx.fee)
-            .ok_or_else(|| SentrixError::InvalidTransaction(
-                "amount + fee overflow".to_string()
-            ))?;
-
-        if available < needed {
-            return Err(SentrixError::InsufficientBalance {
-                have: available,
-                need: needed,
-            });
-        }
-
-        // Insert sorted by fee descending (highest fee = front of queue)
-        // V5-07 TODO: RBF (Replace-By-Fee) not yet implemented — a sender cannot replace
-        // a pending tx with a higher-fee version. Adding RBF requires nonce-keyed lookup
-        // and per-sender replacement logic. Track in BIBLE.md under "Future Work".
-        let pos = self.mempool.iter()
-            .position(|existing| existing.fee < tx.fee)
-            .unwrap_or(self.mempool.len());
-        self.mempool.insert(pos, tx);
-        Ok(())
-    }
-
-    fn mempool_pending_count(&self, address: &str) -> u64 {
-        self.mempool.iter()
-            .filter(|tx| tx.from_address == address)
-            .count() as u64
-    }
-
-    fn mempool_pending_spend(&self, address: &str) -> u64 {
-        self.mempool.iter()
-            .filter(|tx| tx.from_address == address)
-            .map(|tx| tx.amount.saturating_add(tx.fee))
-            .fold(0u64, |acc, v| acc.saturating_add(v))
-    }
-
-    pub fn mempool_size(&self) -> usize {
-        self.mempool.len()
-    }
-
-    /// M-04 FIX: Remove transactions older than MEMPOOL_MAX_AGE_SECS.
-    /// Called automatically after each block is added; also callable manually.
-    pub fn prune_mempool(&mut self) {
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        self.mempool.retain(|tx| now <= tx.timestamp.saturating_add(MEMPOOL_MAX_AGE_SECS));
-    }
-
     // ── Chain validation ─────────────────────────────────
     pub fn is_valid_chain(&self) -> bool {
         for i in 1..self.chain.len() {
@@ -270,537 +162,12 @@ impl Blockchain {
         true
     }
 
-    // ── Block creation (validator calls this) ────────────
-    pub fn create_block(&mut self, validator_address: &str) -> SentrixResult<Block> {
-        let next_height = self.height() + 1;
-
-        // Check authorization
-        if !self.authority.is_authorized(validator_address, next_height)? {
-            return Err(SentrixError::NotYourTurn);
-        }
-
-        // Build transaction list — coinbase first
-        let reward = self.get_block_reward();
-        let coinbase = Transaction::new_coinbase(
-            validator_address.to_string(),
-            reward,
-            next_height,
-        );
-
-        let mut transactions = vec![coinbase];
-
-        // Take up to MAX_TX_PER_BLOCK from mempool
-        let take = self.mempool.len().min(MAX_TX_PER_BLOCK - 1);
-        let mempool_txs: Vec<Transaction> = self.mempool.drain(..take).collect();
-        transactions.extend(mempool_txs);
-
-        let block = Block::new(
-            next_height,
-            self.latest_block()?.hash.clone(),
-            transactions,
-            validator_address.to_string(),
-        );
-
-        Ok(block)
-    }
-
     // I-01 FIX: memory estimate now reflects window (not full chain)
     pub fn get_memory_estimate(&self) -> String {
         let window_blocks = self.chain.len();
         let true_height = self.height();
         let estimate_mb = (window_blocks * 2) / 1024; // ~2KB per block
         format!("~{}MB for {} blocks in window (true height: {})", estimate_mb, window_blocks, true_height)
-    }
-
-    // ── Block application (two-pass atomic) ─────────────
-    pub fn add_block(&mut self, block: Block) -> SentrixResult<()> {
-        let expected_index = self.height() + 1;
-        let expected_prev = self.latest_block()?.hash.clone();
-
-        // ── Pass 1: dry-run validation ───────────────────
-        block.validate_structure(expected_index, &expected_prev)?;
-
-        // C-02 FIX: Verify validator authorization for this block height
-        if !self.authority.is_authorized(&block.validator, expected_index)? {
-            return Err(SentrixError::UnauthorizedValidator(
-                format!("validator {} not authorized for block {}", block.validator, expected_index)
-            ));
-        }
-
-        // H-06 FIX: Validate block timestamp
-        let prev_timestamp = self.latest_block()?.timestamp;
-        let now = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-        if block.timestamp < prev_timestamp {
-            return Err(SentrixError::InvalidBlock(
-                "block timestamp is before previous block".to_string()
-            ));
-        }
-        if block.timestamp > now + 15 {
-            return Err(SentrixError::InvalidBlock(
-                "block timestamp too far in the future".to_string()
-            ));
-        }
-
-        // Validate coinbase amount
-        let reward = self.get_block_reward();
-        let coinbase = block.coinbase()
-            .ok_or_else(|| SentrixError::InvalidBlock("missing coinbase".to_string()))?;
-        if coinbase.amount > reward {
-            return Err(SentrixError::InvalidBlock(
-                format!("coinbase {} exceeds reward {}", coinbase.amount, reward)
-            ));
-        }
-
-        // Validate all non-coinbase transactions on working state copy
-        let mut working_balances: std::collections::HashMap<String, u64> = std::collections::HashMap::new();
-        let mut working_nonces: std::collections::HashMap<String, u64> = std::collections::HashMap::new();
-
-        for tx in block.transactions.iter().skip(1) {
-            // Get working balance (fall back to real balance)
-            let balance = working_balances
-                .get(&tx.from_address)
-                .copied()
-                .unwrap_or_else(|| self.accounts.get_balance(&tx.from_address));
-
-            // Get working nonce
-            let nonce = working_nonces
-                .get(&tx.from_address)
-                .copied()
-                .unwrap_or_else(|| self.accounts.get_nonce(&tx.from_address));
-
-            // Validate
-            tx.validate(nonce, self.chain_id)?;
-
-            // H-02 FIX: checked addition to prevent overflow
-            let needed = tx.amount.checked_add(tx.fee)
-                .ok_or_else(|| SentrixError::InvalidTransaction(
-                    "amount + fee overflow".to_string()
-                ))?;
-            if balance < needed {
-                return Err(SentrixError::InsufficientBalance {
-                    have: balance,
-                    need: needed,
-                });
-            }
-
-            // Validate token operation if present
-            if let Some(token_op) = TokenOp::decode(&tx.data) {
-                match &token_op {
-                    TokenOp::Transfer { contract, amount, .. } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(
-                                format!("token contract {} not found", contract)
-                            ));
-                        }
-                        let token_bal = self.contracts.get_token_balance(contract, &tx.from_address);
-                        if token_bal < *amount {
-                            return Err(SentrixError::InsufficientBalance { have: token_bal, need: *amount });
-                        }
-                    }
-                    TokenOp::Burn { contract, amount } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(
-                                format!("token contract {} not found", contract)
-                            ));
-                        }
-                        let token_bal = self.contracts.get_token_balance(contract, &tx.from_address);
-                        if token_bal < *amount {
-                            return Err(SentrixError::InsufficientBalance { have: token_bal, need: *amount });
-                        }
-                    }
-                    TokenOp::Mint { contract, .. } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(
-                                format!("token contract {} not found", contract)
-                            ));
-                        }
-                    }
-                    TokenOp::Approve { contract, .. } => {
-                        if !self.contracts.exists(contract) {
-                            return Err(SentrixError::InvalidTransaction(
-                                format!("token contract {} not found", contract)
-                            ));
-                        }
-                    }
-                    TokenOp::Deploy { .. } => {} // deploy creates new contract, no pre-check needed
-                }
-            }
-
-            // Update working state
-            *working_balances.entry(tx.from_address.clone()).or_insert(balance) -= needed;
-            *working_nonces.entry(tx.from_address.clone()).or_insert(nonce) += 1;
-        }
-
-        // ── Pass 2: commit ───────────────────────────────
-        // Apply coinbase reward
-        self.accounts.credit(&block.validator, coinbase.amount)?;
-        self.total_minted += coinbase.amount;
-
-        // Apply all transactions
-        let mut total_fee: u64 = 0;
-        for tx in block.transactions.iter().skip(1) {
-            self.accounts.transfer(
-                &tx.from_address,
-                &tx.to_address,
-                tx.amount,
-                tx.fee,
-            )?;
-            total_fee += tx.fee;
-
-            // Execute token operation if present in data field
-            if let Some(token_op) = TokenOp::decode(&tx.data) {
-                match token_op {
-                    TokenOp::Deploy { name, symbol, decimals, supply, max_supply } => {
-                        self.contracts.deploy(&tx.from_address, &name, &symbol, decimals, supply, max_supply)?;
-                    }
-                    TokenOp::Transfer { contract, to, amount } => {
-                        self.contracts.execute_transfer(&contract, &tx.from_address, &to, amount)?;
-                    }
-                    TokenOp::Burn { contract, amount } => {
-                        self.contracts.execute_burn(&contract, &tx.from_address, amount)?;
-                    }
-                    TokenOp::Mint { contract, to, amount } => {
-                        self.contracts.execute_mint(&contract, &tx.from_address, &to, amount)?;
-                    }
-                    TokenOp::Approve { contract, spender, amount } => {
-                        self.contracts.execute_approve(&contract, &tx.from_address, &spender, amount)?;
-                    }
-                }
-            }
-        }
-
-        // L-02 FIX: Burn gets ceiling, validator gets floor — ensures total_fee is fully distributed
-        let burn_fee_share = (total_fee + 1) / 2;
-        let validator_fee_share = total_fee - burn_fee_share;
-        if validator_fee_share > 0 {
-            self.accounts.credit(&block.validator, validator_fee_share)?;
-        }
-
-        // Record validator stats
-        self.authority.record_block_produced(&block.validator, block.timestamp);
-
-        // Remove mined transactions from mempool
-        let mined_txids: std::collections::HashSet<String> = block.transactions
-            .iter()
-            .map(|tx| tx.txid.clone())
-            .collect();
-        self.mempool.retain(|tx| !mined_txids.contains(&tx.txid));
-
-        // M-04 FIX: Prune stale transactions after each block
-        self.prune_mempool();
-
-        // Append block to chain
-        self.chain.push(block);
-
-        // I-01 FIX: sliding window — evict oldest blocks that exceed CHAIN_WINDOW_SIZE
-        // Evicted blocks remain in sled storage; only the in-memory window shrinks
-        if self.chain.len() > CHAIN_WINDOW_SIZE {
-            let excess = self.chain.len() - CHAIN_WINDOW_SIZE;
-            self.chain.drain(..excess);
-        }
-
-        Ok(())
-    }
-
-    // ── SRX-20 Token Operations ────────────────────────────
-
-    pub fn deploy_token(
-        &mut self,
-        deployer: &str,
-        name: String,
-        symbol: String,
-        decimals: u8,
-        total_supply: u64,
-        max_supply: u64,
-        deploy_fee: u64,
-    ) -> SentrixResult<String> {
-        // Check deployer has enough for fee
-        let balance = self.accounts.get_balance(deployer);
-        if balance < deploy_fee {
-            return Err(SentrixError::InsufficientBalance {
-                have: balance,
-                need: deploy_fee,
-            });
-        }
-
-        // Deduct fee: 50% burned, 50% to ecosystem fund
-        if deploy_fee > 0 {
-            let deployer_acc = self.accounts.get_or_create(deployer);
-            deployer_acc.balance = deployer_acc.balance.checked_sub(deploy_fee)
-                .ok_or_else(|| SentrixError::Internal("deploy fee underflow".to_string()))?;
-
-            let burn_share = (deploy_fee + 1) / 2; // L-02 FIX: burn rounds up
-            let eco_share = deploy_fee - burn_share;
-            self.accounts.total_burned += burn_share;
-            self.accounts.credit(ECOSYSTEM_FUND_ADDRESS, eco_share)?;
-        }
-
-        // Deploy contract
-        let contract_address = self.contracts.deploy(deployer, &name, &symbol, decimals, total_supply, max_supply)?;
-        Ok(contract_address)
-    }
-
-    pub fn token_transfer(
-        &mut self,
-        contract_address: &str,
-        caller: &str,
-        to: &str,
-        amount: u64,
-        gas_fee: u64,
-    ) -> SentrixResult<()> {
-        // Check caller has enough SRX for gas
-        let balance = self.accounts.get_balance(caller);
-        if balance < gas_fee {
-            return Err(SentrixError::InsufficientBalance {
-                have: balance,
-                need: gas_fee,
-            });
-        }
-
-        // Deduct gas: 50% burned, 50% to current validator (or ecosystem fund)
-        if gas_fee > 0 {
-            let caller_acc = self.accounts.get_or_create(caller);
-            caller_acc.balance = caller_acc.balance.checked_sub(gas_fee)
-                .ok_or_else(|| SentrixError::Internal("gas fee underflow".to_string()))?;
-
-            let burn_share = (gas_fee + 1) / 2; // L-02 FIX: burn rounds up
-            let val_share = gas_fee - burn_share;
-            self.accounts.total_burned += burn_share;
-
-            let validator_addr = self.authority
-                .expected_validator(self.height() + 1)
-                .map(|v| v.address.clone())
-                .unwrap_or_else(|_| ECOSYSTEM_FUND_ADDRESS.to_string());
-            self.accounts.credit(&validator_addr, val_share)?;
-        }
-
-        // Execute token transfer
-        let contract = self.contracts.get_contract_mut(contract_address)
-            .ok_or_else(|| SentrixError::NotFound(format!("contract {}", contract_address)))?;
-        contract.transfer(caller, to, amount)?;
-        Ok(())
-    }
-
-    pub fn token_burn(
-        &mut self,
-        contract_address: &str,
-        caller: &str,
-        amount: u64,
-        gas_fee: u64,
-    ) -> SentrixResult<()> {
-        // Check caller has enough SRX for gas
-        let balance = self.accounts.get_balance(caller);
-        if balance < gas_fee {
-            return Err(SentrixError::InsufficientBalance {
-                have: balance,
-                need: gas_fee,
-            });
-        }
-
-        // Deduct gas: 50% burned, 50% to validator
-        if gas_fee > 0 {
-            let caller_acc = self.accounts.get_or_create(caller);
-            caller_acc.balance = caller_acc.balance.checked_sub(gas_fee)
-                .ok_or_else(|| SentrixError::Internal("gas fee underflow".to_string()))?;
-
-            let burn_share = (gas_fee + 1) / 2; // L-02 FIX: burn rounds up
-            let val_share = gas_fee - burn_share;
-            self.accounts.total_burned += burn_share;
-
-            let validator_addr = self.authority
-                .expected_validator(self.height() + 1)
-                .map(|v| v.address.clone())
-                .unwrap_or_else(|_| ECOSYSTEM_FUND_ADDRESS.to_string());
-            self.accounts.credit(&validator_addr, val_share)?;
-        }
-
-        // Execute token burn
-        let contract = self.contracts.get_contract_mut(contract_address)
-            .ok_or_else(|| SentrixError::NotFound(format!("contract {}", contract_address)))?;
-        contract.burn(caller, amount)?;
-        Ok(())
-    }
-
-    pub fn token_balance(&self, contract_address: &str, address: &str) -> u64 {
-        self.contracts.get_contract(contract_address)
-            .map(|c| c.balance_of(address))
-            .unwrap_or(0)
-    }
-
-    pub fn token_info(&self, contract_address: &str) -> SentrixResult<serde_json::Value> {
-        let contract = self.contracts.get_contract(contract_address)
-            .ok_or_else(|| SentrixError::NotFound(format!("contract {}", contract_address)))?;
-        Ok(contract.get_info())
-    }
-
-    pub fn list_tokens(&self) -> Vec<serde_json::Value> {
-        self.contracts.list_contracts()
-    }
-
-    // ── Transaction queries ──────────────────────────────
-
-    pub fn get_transaction(&self, txid: &str) -> Option<serde_json::Value> {
-        for block in self.chain.iter().rev() {
-            for tx in &block.transactions {
-                if tx.txid == txid {
-                    return Some(serde_json::json!({
-                        "transaction": tx,
-                        "block_index": block.index,
-                        "block_hash": block.hash,
-                        "block_timestamp": block.timestamp,
-                    }));
-                }
-            }
-        }
-        None
-    }
-
-    // L-03 FIX: paginated address history (limit + offset)
-    pub fn get_address_history(&self, address: &str, limit: usize, offset: usize) -> Vec<serde_json::Value> {
-        let mut history = Vec::new();
-        let mut skipped = 0usize;
-        for block in &self.chain {
-            for tx in &block.transactions {
-                if tx.from_address == address || tx.to_address == address {
-                    if skipped < offset {
-                        skipped += 1;
-                        continue;
-                    }
-                    if history.len() >= limit {
-                        return history;
-                    }
-                    let direction = if tx.from_address == address {
-                        if tx.is_coinbase() { "reward" } else { "out" }
-                    } else {
-                        "in"
-                    };
-                    history.push(serde_json::json!({
-                        "txid": tx.txid,
-                        "direction": direction,
-                        "from": tx.from_address,
-                        "to": tx.to_address,
-                        "amount": tx.amount,
-                        "fee": tx.fee,
-                        "block_index": block.index,
-                        "block_timestamp": block.timestamp,
-                    }));
-                }
-            }
-        }
-        history
-    }
-
-    pub fn get_address_tx_count(&self, address: &str) -> usize {
-        self.chain.iter()
-            .flat_map(|b| &b.transactions)
-            .filter(|tx| tx.from_address == address || tx.to_address == address)
-            .count()
-    }
-
-    pub fn get_latest_transactions(&self, limit: usize, offset: usize) -> Vec<serde_json::Value> {
-        let mut result = Vec::new();
-        let mut skipped = 0usize;
-        for block in self.chain.iter().rev() {
-            for tx in block.transactions.iter().rev() {
-                if skipped < offset {
-                    skipped += 1;
-                    continue;
-                }
-                if result.len() >= limit {
-                    return result;
-                }
-                result.push(serde_json::json!({
-                    "txid": tx.txid,
-                    "from": tx.from_address,
-                    "to": tx.to_address,
-                    "amount": tx.amount,
-                    "fee": tx.fee,
-                    "is_coinbase": tx.is_coinbase(),
-                    "block_index": block.index,
-                    "block_timestamp": block.timestamp,
-                }));
-            }
-        }
-        result
-    }
-
-    pub fn get_token_holders(&self, contract: &str) -> Option<Vec<serde_json::Value>> {
-        self.contracts.get_holders_list(contract)
-    }
-
-    pub fn get_token_trades(&self, contract_addr: &str, limit: usize, offset: usize) -> Vec<serde_json::Value> {
-        let mut result = Vec::new();
-        let mut skipped = 0usize;
-        for block in self.chain.iter().rev() {
-            for tx in block.transactions.iter() {
-                let entry = match TokenOp::decode(&tx.data) {
-                    Some(TokenOp::Transfer { contract, to, amount }) if contract == contract_addr => {
-                        Some(serde_json::json!({
-                            "type": "transfer",
-                            "from": tx.from_address,
-                            "to": to,
-                            "amount": amount,
-                            "txid": tx.txid,
-                            "block_index": block.index,
-                            "block_timestamp": block.timestamp,
-                        }))
-                    }
-                    Some(TokenOp::Burn { contract, amount }) if contract == contract_addr => {
-                        Some(serde_json::json!({
-                            "type": "burn",
-                            "from": tx.from_address,
-                            "to": serde_json::Value::Null,
-                            "amount": amount,
-                            "txid": tx.txid,
-                            "block_index": block.index,
-                            "block_timestamp": block.timestamp,
-                        }))
-                    }
-                    Some(TokenOp::Mint { contract, to, amount }) if contract == contract_addr => {
-                        Some(serde_json::json!({
-                            "type": "mint",
-                            "from": tx.from_address,
-                            "to": to,
-                            "amount": amount,
-                            "txid": tx.txid,
-                            "block_index": block.index,
-                            "block_timestamp": block.timestamp,
-                        }))
-                    }
-                    _ => None,
-                };
-                if let Some(e) = entry {
-                    if skipped < offset {
-                        skipped += 1;
-                    } else {
-                        result.push(e);
-                        if result.len() >= limit {
-                            return result;
-                        }
-                    }
-                }
-            }
-        }
-        result
-    }
-
-    // ── Stats ────────────────────────────────────────────
-    pub fn chain_stats(&self) -> serde_json::Value {
-        serde_json::json!({
-            "height": self.height(),
-            "total_blocks": self.height() + 1, // I-01 FIX: chain window may be < total blocks
-            "total_minted_srx": self.total_minted as f64 / 100_000_000.0,
-            "max_supply_srx": MAX_SUPPLY as f64 / 100_000_000.0,
-            "total_burned_srx": self.accounts.total_burned as f64 / 100_000_000.0,
-            "mempool_size": self.mempool.len(),
-            "active_validators": self.authority.active_count(),
-            "deployed_tokens": self.contracts.contract_count(),
-            "chain_id": self.chain_id,
-            "next_block_reward_srx": self.get_block_reward() as f64 / 100_000_000.0,
-        })
     }
 }
