+++ src/core/mod.rs
@@ -3,6 +3,11 @@ pub mod block;
 pub mod transaction;
 pub mod account;
 pub mod blockchain;
+pub mod mempool;
+pub mod block_producer;
+pub mod block_executor;
+pub mod token_ops;
+pub mod chain_queries;
 pub mod authority;
 pub mod merkle;
 pub mod vm;
